extract ClientCredentialBase
